Refresh project
So many of the project's original goals need to be changed. Thanks to
the whole Camridge Analytica issue, Facebook changed A LOT of things
about their Graph API overnight, most of them being breaking changes.
The most relevant change being that no personal data can be retrieved
even from posts even on public Facebook groups. So, it is for now
impossible to know who made a post, who liked it, who commented on it,
etc. Update the project to Facebook's 3.0 version of the Graph API
and hard code personal information for now, keeping it as a placeholder
for the duration before Facebook (if ever) reverts back their changes.